services/chat-api/requirements.txt
fastapi
uvicorn
openai
python-dotenv
qdrant-client
pypdf
python-docx
python-multipart
tiktoken
tenacity
python-jose[cryptography]
